Revert "test(e2e): borrow #3317's blob fix to prove the whole chain — REVERT BEFORE MERGE"
This reverts commit dee95d271c2135fe2dade3d495edccd41fa37cce.
