+++ crates/store/src/lib.rs
@@ -11,8 +11,6 @@ pub mod handle;
 pub mod iter;
 pub mod key;
 pub mod layer;
-#[cfg(feature = "datatypes")]
-pub mod namespace_signer;
 pub mod slice;
 pub mod tx;
